+++ benchmarks/dynamo/torchbench.py
@@ -1,13 +1,8 @@
 #!/usr/bin/env python3
 
-import gc
-import importlib
 import logging
 import os
-import re
-import sys
 import warnings
-from collections import namedtuple
 
 import torch
 
@@ -20,20 +15,26 @@ except ImportError:
 from torch._dynamo.testing import collect_results, reduce_to_scalar_loss
 from torch._dynamo.utils import clone_inputs
 
+
 try:
     from .torchbench_output_processors import PROCESS_TRAIN_MODEL_OUTPUT
 except ImportError:
     from torchbench_output_processors import PROCESS_TRAIN_MODEL_OUTPUT
 
 try:
-    from .torchbench_utils import _reassign_parameters, setup_torchbench_cwd
+    from .torchbench_utils import setup_torchbench_cwd
 except ImportError:
-    from torchbench_utils import _reassign_parameters, setup_torchbench_cwd
+    from torchbench_utils import setup_torchbench_cwd
 
 try:
     from .torchbench_config import TorchBenchConfig
 except ImportError:
     from torchbench_config import TorchBenchConfig
+
+try:
+    from .torchbench_model_loader import TorchBenchModelLoader
+except ImportError:
+    from torchbench_model_loader import TorchBenchModelLoader
 
 
 # We are primarily interested in tf32 datatype
@@ -48,21 +49,19 @@ if "TORCHINDUCTOR_AUTOGRAD_CACHE" not in os.environ:
     torch._functorch.config.enable_autograd_cache = True
 
 
-
-
-
-
 class TorchBenchmarkRunner(BenchmarkRunner):
     def __init__(self):
         super().__init__()
         self.suite_name = "torchbench"
         self.optimizer = None
         self._tb_config = TorchBenchConfig()
+        self._model_loader = None
 
     def _ensure_config_setup(self):
         """Ensure config is setup with args if not already done."""
-        if hasattr(self, 'args') and not hasattr(self._tb_config, 'args'):
+        if hasattr(self, "args") and not hasattr(self._tb_config, "args"):
             self._tb_config.set_args(self.args)
+            self._model_loader = TorchBenchModelLoader(self._tb_config, self.args)
 
     # Delegate configuration properties to the config object
     @property
@@ -159,164 +158,33 @@ class TorchBenchmarkRunner(BenchmarkRunner):
         part=None,
         extra_args=None,
     ):
-        if self.args.enable_activation_checkpointing:
-            raise NotImplementedError(
-                "Activation checkpointing not implemented for Torchbench models"
-            )
-        is_training = self.args.training
-        use_eval_mode = self.args.use_eval_mode
-        candidates = [
-            f"torchbenchmark.models.{model_name}",
-            f"torchbenchmark.canary_models.{model_name}",
-            f"torchbenchmark.models.fb.{model_name}",
-        ]
-        for c in candidates:
-            try:
-                module = importlib.import_module(c)
-                break
-            except ModuleNotFoundError as e:
-                if e.name != c:
-                    raise
-        else:
-            raise ImportError(f"could not import any of {candidates}")
-        benchmark_cls = getattr(module, "Model", None)
-        if benchmark_cls is None:
-            raise NotImplementedError(f"{model_name}.Model is None")
-
-        if not hasattr(benchmark_cls, "name"):
-            benchmark_cls.name = model_name
-
-        cant_change_batch_size = (
-            not getattr(benchmark_cls, "ALLOW_CUSTOMIZE_BSIZE", True)
-            or model_name in self._tb_config._config["dont_change_batch_size"]
-        )
-        if cant_change_batch_size:
-            batch_size = None
-        
         self._ensure_config_setup()
-        batch_size = self._tb_config.get_batch_size_for_model(model_name, is_training, batch_size)
-        
-        # Control the memory footprint for few models
-        batch_size = self._tb_config.limit_batch_size_for_accuracy(model_name, batch_size)
 
-        # workaround "RuntimeError: not allowed to set torch.backends.cudnn flags"
-        torch.backends.__allow_nonbracketed_mutation_flag = True
-        if extra_args is None:
-            extra_args = []
-        if part:
-            extra_args += ["--part", part]
-
-        # sam_fast only runs with amp
+        # Handle sam_fast special case that needs setup_amp
         if model_name == "sam_fast":
             self.args.amp = True
             self.setup_amp()
 
-        if model_name == "vision_maskrcnn" and is_training:
-            # Output of vision_maskrcnn model is a list of bounding boxes,
-            # sorted on the basis of their scores. This makes accuracy
-            # comparison hard with torch.compile. torch.compile can cause minor
-            # divergences in the output because of how fusion works for amp in
-            # TorchInductor compared to eager.  Therefore, instead of looking at
-            # all the bounding boxes, we compare only top 4.
-            model_kwargs = {"box_detections_per_img": 4}
-            benchmark = benchmark_cls(
-                test="train",
-                device=device,
-                batch_size=batch_size,
-                extra_args=extra_args,
-                model_kwargs=model_kwargs,
+        device, model_name, model, example_inputs, batch_size = (
+            self._model_loader.load_model(
+                device, model_name, batch_size, part, extra_args
             )
-            use_eval_mode = True
-        elif is_training:
-            benchmark = benchmark_cls(
-                test="train",
-                device=device,
-                batch_size=batch_size,
-                extra_args=extra_args,
-            )
-        else:
-            benchmark = benchmark_cls(
-                test="eval",
-                device=device,
-                batch_size=batch_size,
-                extra_args=extra_args,
-            )
-        model, example_inputs = benchmark.get_module()
-        if model_name in [
-            "basic_gnn_edgecnn",
-            "basic_gnn_gcn",
-            "basic_gnn_sage",
-            "basic_gnn_gin",
-        ]:
-            _reassign_parameters(model)
+        )
 
-        # Models that must be in train mode while training
-        if is_training and (
-            not use_eval_mode or model_name in self._tb_config._config["only_training"]
-        ):
-            model.train()
-        else:
-            model.eval()
-        gc.collect()
-        batch_size = benchmark.batch_size
-        if model_name == "torchrec_dlrm":
-            batch_namedtuple = namedtuple(
-                "Batch", "dense_features sparse_features labels"
-            )
-            example_inputs = tuple(
-                batch_namedtuple(
-                    dense_features=batch.dense_features,
-                    sparse_features=batch.sparse_features,
-                    labels=batch.labels,
-                )
-                for batch in example_inputs
-            )
-        # Torchbench has quite different setup for yolov3, so directly passing
-        # the right example_inputs
-        if model_name == "yolov3":
-            example_inputs = (torch.rand(batch_size, 3, 384, 512).to(device),)
-        # See https://github.com/pytorch/benchmark/issues/1561
-        if model_name == "maml_omniglot":
-            batch_size = 5
-            assert example_inputs[0].shape[0] == batch_size
+        # Handle special batch size overrides that the model loader doesn't handle
         if model_name == "vision_maskrcnn":
             batch_size = 1
-        # global current_name, current_device
-        # current_device = device
-        # current_name = benchmark.name
 
-        if self.args.trace_on_xla:
-            # work around for: https://github.com/pytorch/xla/issues/4174
-            import torch_xla  # noqa: F401
         self.validate_model(model, example_inputs)
-        return device, benchmark.name, model, example_inputs, batch_size
+        return device, model_name, model, example_inputs, batch_size
 
     def iter_model_names(self, args):
-        from torchbenchmark import _list_canary_model_paths, _list_model_paths
-
-        models = _list_model_paths()
-        models += [
-            f
-            for f in _list_canary_model_paths()
-            if os.path.basename(f) in self._tb_config._config["canary_models"]
-        ]
-        models.sort()
-
-        start, end = self.get_benchmark_indices(len(models))
-        for index, model_path in enumerate(models):
-            if index < start or index >= end:
-                continue
-
-            model_name = os.path.basename(model_path)
-            if (
-                not re.search("|".join(args.filter), model_name, re.IGNORECASE)
-                or re.search("|".join(args.exclude), model_name, re.IGNORECASE)
-                or model_name in args.exclude_exact
-                or model_name in self.skip_models
-            ):
-                continue
-
-            yield model_name
+        self._ensure_config_setup()
+        # Update the model loader to use the right benchmark indices method
+        self._model_loader._get_benchmark_indices = (
+            lambda total: self.get_benchmark_indices(total)
+        )
+        return self._model_loader.iter_model_names()
 
     def pick_grad(self, name, is_training):
         if is_training or name in ("maml",):
@@ -329,7 +197,9 @@ class TorchBenchmarkRunner(BenchmarkRunner):
 
     def get_tolerance_and_cosine_flag(self, is_training, current_device, name):
         self._ensure_config_setup()
-        return self._tb_config.get_tolerance_and_cosine_flag(is_training, current_device, name, self.args)
+        return self._tb_config.get_tolerance_and_cosine_flag(
+            is_training, current_device, name, self.args
+        )
 
     def compute_loss(self, pred):
         return reduce_to_scalar_loss(pred)
